% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankOMOPTables.R
\name{blankOMOPTables}
\alias{blankOMOPTables}
\title{blankOMOPTables}
\usage{
blankOMOPTables(definition_table, table_names = NULL)
}
\arguments{
\item{definition_table}{Data file with definitions of OMOP terms and their corresponding names in the standard file.}

\item{table_names}{Names of tables to return, if any. If NULL, returns all tables.}
}
\description{
Function uses the specified table_defintions file to generate empty tables
with formatted columns. If table names are provided, only returns tables
with the provided names.
}
