% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankStandardTable.R
\name{blankStandardTable}
\alias{blankStandardTable}
\title{blankStandardTable}
\usage{
blankStandardTable()
}
\description{
Function returns a blank standard table as a template.
}
\details{
blankStandardTable()
}
