% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSQLDB.R
\name{buildSQLDB}
\alias{buildSQLDB}
\title{buildSQLDBR.R}
\usage{
buildSQLDB(
  omop_csvs,
  sql_db_file,
  csv_to_sqlite_loc = NULL,
  return_cmd_only = FALSE
)
}
\arguments{
\item{omop_csvs}{Filesnames of all OMOP csv files to be incorporated into the database.}

\item{sql_db_file}{Filename under which to store the SQLite database file.}

\item{csv_to_sqlite_loc}{Location of csv-to-sqlite executable. If not provided, function attempts to use executable in the path.}

\item{return_cmd_only}{For debugging purposes; if TRUE, returns only the command to be executed and does not execute it.

buildSQLDB()}
}
\description{
Function uses the python csv-to-sqlite function to create a SQLite database.
Requires that csv-to-sqlite is in the path, or provided via the
<csv_to_sqlite_loc> argument. Returns the filename of the database created.
}
