% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSQLDBR.R
\name{buildSQLDBR}
\alias{buildSQLDBR}
\title{buildSQLDBR}
\usage{
buildSQLDBR(omop_tables, sql_db_file)
}
\arguments{
\item{omop_tables}{Filesnames of all OMOP csv files to be incorporated into the database.}

\item{sql_db_file}{Filename under which to store the SQLite database file.}
}
\description{
Wrapper which uses dplyr's src_sqlite() function to create the SQLite 
database, and adds all OMOP-formatted tables (as produced by the function
combineInputTables()) to the database. Requires a list of formatted tables
and an output file name within which to store the SQL database.
}
