% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geo_by_url.R
\name{check_geo_by_url}
\alias{check_geo_by_url}
\title{check_geo_by_url}
\usage{
check_geo_by_url(id_input = "GSE15896")
}
\arguments{
\item{id_input}{Character string of GSE or GDS ID; i.e. "GSE15896".}
}
\description{
Background function (not user-facing).
}
\details{
Given a GSE or GDS ID, function builds a corresponding URL and tests it to 
ensure the ID is valid. Returns boolean results of "url.exists()" from RCURL
pacakge.
}
