% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileID.R
\name{compileID}
\alias{compileID}
\title{compileID}
\usage{
compileID(table_input, id_column, id_pfx, inc_columns)
}
\arguments{
\item{table_input}{Standard table being processed.}

\item{id_column}{String with the intended name of the new ID column.}

\item{id_pfx}{If a prefix is to be appended to the ID, include it here.}

\item{inc_columns}{Names of existing columns in the table to be included.}
}
\description{
Given an input standard table and the names of each column to be included in
an OMOP table, function adds a column named <id_column> with a unique ID for
each permutation of the other included columns. Columns not found in the
table are ignored without a warning.
}
