% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_geo_table.R
\name{composite_geo_table}
\alias{composite_geo_table}
\title{composite_geo_table}
\usage{
composite_geo_table(geo_input = NULL)
}
\arguments{
\item{geo_input}{GEO input objet (list) from a fetch function.}
}
\description{
Background function (not user-facing).
}
\details{
Given a GEO list object from one of the fetch functions, composite_geo_table
builds a table of metadata based on metadata from comprised GDS, GSE, and 
GPL objects.
}
