% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_and_tpose.R
\name{expand_and_tpose}
\alias{expand_and_tpose}
\title{expand_and_tpose}
\usage{
expand_and_tpose(input_frame, rows_out = 10, name_col = NULL, vals_col = NULL)
}
\arguments{
\item{input_frame}{Dataframe to be transposed.}

\item{rows_out}{Number of rows to expand by.}

\item{name_col}{Name of column containing transposed column names. Defaults to first column.}

\item{vals_col}{Name of column containing transposed column values. Defaults to second column.}
}
\description{
Background function (not user-facing).
}
\details{
Given a dataframe with at least two columns, transpose it using one column 
for names and another for values (defaults to 1 and 2, respectively) and 
duplicate into as many rows as needed.
}
