% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_geo_series.R
\name{fetch_geo_series}
\alias{fetch_geo_series}
\title{fetch_geo_series}
\usage{
fetch_geo_series(geo_series_ids, data_dir = NULL)
}
\arguments{
\item{geo_series_ids}{GEO series ID (e.g. "GSE158946") or a list of such IDs.}

\item{data_dir}{Directory into which retrieved data objects should be stored. NULL by default, in which case saved to memory in a temp folder.}
}
\description{
Wrapper for most GEO import functions when incorporating GEOquery sample
series'. Given a GEO series ID (i.e. "GSE158946") or list of such IDs,
function retrieves S4 metadata obejcts from GEOquery for each dataset, as 
well as its associated platform (GPL) and column data. Function also 
produces a metadata table for each such component, and stores these within
nested list. A comprehensive metadata table is also produced and stored. If
multiple IDs are submitted, these are imported recursively, and metadata 
from each is a merged into a comprehensive metadata table in the top-level
list. Invalid ID entries are identified by checking the validity of their 
purported URL: if invalid, a warning message is displayed. With multiple
IDs, these are added to a text object denoting which failed.
}
