% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderConceptID.R
\name{genderConceptID}
\alias{genderConceptID}
\title{genderConceptID}
\usage{
genderConceptID(gender_val)
}
\arguments{
\item{gender_val}{Accepts male/female/NA, M/F/NA, 8507/8532/8551.

genderConceptID()}
}
\description{
Function to translate assorted male/female/NA inputs into gender concept IDs.
}
