% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBTables.R
\name{getDBTables}
\alias{getDBTables}
\title{getDBTables}
\usage{
getDBTables(database_object, find_key = TRUE, quiet = FALSE)
}
\arguments{
\item{database_object}{A built SQLite database object, typically output by dbOMOP().}

\item{find_key}{Boolean; if TRUE, getTableDependencies() is used to identify and sort tables by dependencies.}

\item{quiet}{Boolean; if TRUE, messages are reported for find_key steps (either names of independent tables or a warning that none were found).

getDBTables()}
}
\description{
Given a database object, function returns all attached tables in a list. If
find_key = TRUE, function calls getTableDependencies() to identify which
tables in the list are dependent on others, and attempts to find at least
one table with no dependencies. If any are found, they are moved to
the front of the list such that later query functions should always include
at least one of them to ensure that a path between tables exists for all
possible queries.
}
