% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDependencies.R
\name{getDependencies}
\alias{getDependencies}
\title{getDependencies}
\usage{
getDependencies(data_model = loadDataModel(), inc_tabs, return_matrix = FALSE)
}
\arguments{
\item{data_model}{Data model from loadDataModel(), in single-table format.}

\item{inc_tabs}{Only include tables with these names.}

\item{return_matrix}{Return dependencies as integer-format matrix of directional connections.

getDependencies()}
}
\description{
Function accepts a data model from loadDataModel() and returns a list of
dependencies between constitutive tables based on the presence of index
columns (columns that end with "<table_name>_id").
}
