% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableDependencies.R
\name{getTableDependencies}
\alias{getTableDependencies}
\title{getTableDependencies}
\usage{
getTableDependencies(table_list)
}
\arguments{
\item{table_list}{Named list of tables pulled from an existing SQLite DB object.

getTableDependencies()}
}
\description{
Given a list of tables from an SQLite DB object, function identifies each
table's dependencies. 'Dependency' here is defined as requiring an index
column from another table. For instance, Table A depends on Table B if it
includes Table B's index column. I.e., Table A cannot be queried without
Table B. For the most part, this is useful for identifying at least one
independent table whose inclusion should guarantee that all queries are
possible, though results could be useful elsewhere.
}
\details{
Function assumes that all table IDs are in the "<table_name>_id" format.
}
