% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankStandardTable.R, R/inputTable.R
\name{inputTable}
\alias{inputTable}
\title{inputTable}
\usage{
inputTable(master_table_file)

inputTable(master_table_file)
}
\description{
Function returns a blank standard table as a template, given an included master table file.

Function returns a blank standard table as a template, given an included master table file.
}
\details{
inputTable()

inputTable()
}
