% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_class.R
\name{interpret_class}
\alias{interpret_class}
\title{interpret_class}
\usage{
interpret_class(omop_def = "VARCHAR(50)")
}
\arguments{
\item{omop_def}{OMOP field definition to interpret.

interpret_class()}
}
\description{
Given a character string from the default OMOP "type" column, convert that
to the tidyverse equivalent ("INTEGER" = "integer", "VARCHAR(x)" =
"character", etc.). As of now, there are no restrictions like character
character limits, etc.
}
