% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelMasks.R
\name{load_mask_dir}
\alias{load_mask_dir}
\title{Load then read mask files from a directory}
\usage{
load_mask_dir(dir_name)
}
\arguments{
\item{dir_name}{Name of directory to search for masks.

load_mask_dir()}
}
\description{
\code{load_mask_dir()} takes a directory, reads in each file (includes only 
mask files which contain the string 'mask'), and approximates a 
a name from the file name (basename, and remove "_mask.[ct]sv"). Throws an 
error if no files are found.
}
