% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelMasks.R
\name{loadModelMasks}
\alias{loadModelMasks}
\title{loadModelMasks}
\usage{
loadModelMasks(mask_files)
}
\arguments{
\item{mask_files}{Either one or more file names of mask TSVs, or a directory of mask TSVs.

loadModelMasks()}
}
\description{
Function reads in all "mask" tables, either within a directory of mask files
ending in ".tsv," a vector of file names of individual mask TSVs, or both.
Automatically adds essential columns table, alias, and field: 'Alias'
denotes the value being input, 'field' denotes the data model's equivalent
field name, and 'table' denotes the data model table where that field
resides. If multiple files are provided, they are returned in a list named
from the basenames of the loaded files. If the list of mask files is already
named, these names are used instead.
}
