% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_mask.R
\name{metadata_mask}
\alias{metadata_mask}
\title{metadata_mask}
\usage{
metadata_mask(metadata_table, examples = 3)
}
\arguments{
\item{metadata_table}{Metadata table to be used.}

\item{examples}{Defaults to three; how many ";" separated examples from the data should be included?

metadata_mask()}
}
\description{
Background function (not typically user-facing).
}
\details{
Function parses a metadata tibble into a mask format, including examples.
}
