% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputOMOPTables.R
\name{outputOMOPTables}
\alias{outputOMOPTables}
\title{outputOMOPTables}
\usage{
outputOMOPTables(standard_table, omop_table_directory)
}
\arguments{
\item{standard_table}{Input standard table with each OMOP table's respective index column added.}

\item{omop_table_directory}{Directory into which OMOP tables will be placed.}
}
\description{
Given a standard table that has been processed to include IDs for all OMOP
tables and an output directory, function summarizes and sorts each into an
output CSV file which it places in the output directory. These CSVs are the
inputs for creating the SQLite database.
}
