% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_metadata.R
\name{parse_geo_metadata}
\alias{parse_geo_metadata}
\title{parse_geo_metadata}
\usage{
parse_geo_metadata(geo_in = NULL)
}
\arguments{
\item{geo_in}{GEO object as returned by getGEO() (GDS or GSE, generally).}
}
\description{
Background function (not typically user-facing).
}
\details{
Breaks metadata into a tibble of all values that apply accross all samples
("all_samples"), i.e. all values that have one entry, and a tibble of all
metadata with multiple values ("other_values"). Also returns a text-based
summary of these "other values" as "other_values_txt".
}
