% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_text_results.R
\name{parse_geo_text_results}
\alias{parse_geo_text_results}
\title{parse_geo_text_results}
\usage{
parse_geo_text_results(
  text_file = file.path(data_dir, "ds_ids_kidney_seqencing.txt")
)
}
\arguments{
\item{text_file}{Text file location containing the "text" format of the GEO browser's results.}
}
\description{
Search results from GEO can be downloaded in text form, and this function 
parses these into a table. More for convinience than anything, this could be
useful if users want to incorporate search results they've generated on the 
GEO website. To use, select the "text" option for the download and save to
a text file, then point this function at that file.
}
