% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gse_metadata.R
\name{parse_gse_metadata}
\alias{parse_gse_metadata}
\title{parse_gse_metadata}
\usage{
parse_gse_metadata(gse_in = NULL)
}
\arguments{
\item{gse_in}{GSE list, as returned by getGEO() (list of ExpressionSet S4 objects).}
}
\description{
Background function (not typically user-facing).
}
\details{
Given a series list (GSE), parse each to generate metadata table, raw file
list, and retrieve the associated abstract.
}
