% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metadata.R
\name{parse_metadata}
\alias{parse_metadata}
\title{parse_metadata}
\usage{
parse_metadata(object_in = NULL)
}
\arguments{
\item{object_in}{GEOquery object returned by getGEO() function.}
}
\description{
Background function (not typically user-facing).
}
\details{
Wrapper function for metadata parsing, since these differ between GDS/GPL 
and GSE types. Basically accepts either object type and passes it to the
appropriate function (parse_geo_metadata() vs. parse_gse_metadata()). Acts 
recursively if a list is provided.
}
