% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseToOMOP.R
\name{parseToOMOP}
\alias{parseToOMOP}
\title{parseToOMOP.R}
\usage{
parseToOMOP(standard_table)
}
\arguments{
\item{standard_table}{Standard table, typically the output of readStandardTables().}
}
\description{
Given a formatted input table, function first renames columns to their
corresponding OMOP term. Next, compileID() is used to append an index column
for each OMOP table to be built.
}
