% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_mask}
\usage{
read_data(file_name)
}
\arguments{
\item{file_name}{File name to be read. Should specify a TSV/CSV file.

read_data()}
}
\description{
Given a data file TSV/CSV, function attempts to read it using fread and returns
an error message if it fails. Returns table with all applicable columns
appended.
}
