% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mask_tsv.R
\name{read_mask_tsv}
\alias{read_mask_tsv}
\title{read_mask_tsv}
\usage{
read_mask_tsv(file_name)
}
\arguments{
\item{file_name}{File name to be read. Should specify a TSV file.

read_mask_tsv()}
}
\description{
Given a mask file TSV, function attempts to read it using fread and returns
an error message if it fails. Returns table with all applicable columns
appended.
}
