% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInputFiles.R
\name{readInputFiles}
\alias{readInputFiles}
\title{readInputFiles}
\usage{
readInputFiles(input_file, data_model, mask_table)
}
\arguments{
\item{input_file}{Name of a TSV file containing required alias column names.}

\item{data_model}{Data model being used, typically as a tibble returned by loadDataModel().}

\item{mask_table}{Mask contained in a tibble, typically as a tibble loaded by loadModelMask().

readInputFiles}
}
\description{
Given a TSV/CSV file containing a data set, that dataset's mask as formatted
by loadModelMask(), and the desired data model produced by loadDataModel(),
this function reads in that data table and funnels its fields into the 
appropriate tables and fields for the data model. All data tables are
included at this stage, including those with no entries, which allows any
data set produced by this function to be incorporated into the same database
as long as the same data model was used. As such, the output tables are
"exhaustive" in that no unused columns are dropped.
}
