% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTableDefinitions.R
\name{standardTableDefinitions}
\alias{standardTableDefinitions}
\title{standardTableDefinitions}
\usage{
standardTableDefinitions(definition_table)
}
\arguments{
\item{definition_table}{Data file with definitions of OMOP terms and their corresponding names in the standard file. If not provided, defaults to stored default.}
}
\description{
Functions reads in stored definition table and returns it as a tibble.
}
