% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInputFile.R
\name{transpose_to_colwise}
\alias{transpose_to_colwise}
\title{switch columns and rows of (transpose) data table}
\usage{
transpose_to_colwise(input_tibble)
}
\arguments{
\item{input_tibble}{Any tibble (no rownames, obviously).}
}
\description{
If an input data set is arranged in a rowwise fashion (one row per patient),
use \code{transpose_to_colwise()} transposes this data.
}
