% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_by_chars.R
\name{truncate_by_chars}
\alias{truncate_by_chars}
\title{truncate_by_chars}
\usage{
truncate_by_chars(text_in, max_char = 80)
}
\arguments{
\item{text_in}{Text string to be truncated.}

\item{max_char}{Maximum number of characters to display (including ellipses). Defaults to 80.

truncate_by_chars}
}
\description{
Background function (not typically user-facing).
}
\details{
Convenience function for displaying things in human-readable tables. After a
given maximum number of characters, cut off the remainder and replace with
an ellipsis.
}
