% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSQLDBR.R
\name{buildSQLDBR}
\alias{buildSQLDBR}
\title{buildSQLDBR}
\usage{
buildSQLDBR(omop_tables, sql_db_file = ":memory:")
}
\arguments{
\item{omop_tables}{Filesnames of all OMOP csv files to be incorporated into the database.}

\item{sql_db_file}{Filename under which to store the SQLite database file.}
}
\description{
Wrapper which uses dplyr's src_sqlite() function to create the SQLite 
database, and adds all OMOP-formatted tables (as produced by the function
combineInputTables()) to the database. Requires a list of formatted tables.
If an an output file name is provided, the database file is saved in this 
location. Otherwise, the database file is stored in memory (equivalent to 
providing ":memory:" to dbConnect()).
}
