% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineInputTables.R
\name{combineInputTables}
\alias{combineInputTables}
\title{combineInputTables}
\usage{
combineInputTables(input_table_list = lapply(seq_input_files, readInputFiles))
}
\arguments{
\item{input_table_list}{List of tables for inclusion, typically from readInputFiles() with different masks but the same base data model.

combineInputTables()}
}
\description{
Given a list of input tables converted into the "exhaustive" format of the
data model using readInputFiles(), convert all of these from all masks into
a single table, eliminate all unused OMOP tables, and index according to
each included table's unique permutations of data. Once complete, return the
data model's definition of tables including formatting for type as a list of
tibbles that can be submitted to buildSQLDBR.R.
}
