% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_entry_columns.R
\name{expand_entry_columns}
\alias{expand_entry_columns}
\title{expand_entry_columns}
\usage{
expand_entry_columns(
  table_in = a,
  col_nms = c("brca_clinical2", "brca_clinical3")
)
}
\arguments{
\item{table_in}{Table to be expanded, generally intermediate format within readInputFiles().}

\item{col_nms}{Names of columns to be "expanded".

expand_entry_columns()}
}
\description{
Function takes in a table with duplicated fields and a field_index column 
(the output of readInputFiles, effectively) and spreads them wider, i.e. 
one column per patient to one column per treatment. Probably more useful
when called by readInputFiles rather than by users.
}
