% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_entry_columns.R
\name{expand_entry_columns}
\alias{expand_entry_columns}
\title{expand_entry_columns}
\usage{
expand_entry_columns(table_in)
}
\arguments{
\item{table_in}{Table to be expanded, generally intermediate format within readInputFiles().

expand_entry_columns()}
}
\description{
Function is similar to dplyr::unnest() on all fields that are used more than
once (assigned a field_idx), but ensures that new columns also contain all 
common field entries (for instance, if the field 'value_as_number' is used 
twice, both indices are given a column, but both columns should feature 
"patient_name" and "gender_source_value".
}
