% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mask_dir.R
\name{load_mask_dir}
\alias{load_mask_dir}
\title{read_mask_dir}
\usage{
load_mask_dir(dir_name)
}
\arguments{
\item{dir_name}{Name of directory to search for masks.

read_mask_dir()}
}
\description{
Given a directory containing mask TSV files, read each and approximate a 
a name from the file name (basename, and remove "_mask.tsv"). Throws an 
error if no files are found.
}
