% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataModel.R
\name{loadDataModel}
\alias{loadDataModel}
\title{loadDataModel}
\usage{
loadDataModel(master_table_file, as_table_list = FALSE)
}
\arguments{
\item{master_table_file}{File containing the total data model used, including 
"field", "required", "type", "description", and "table" fields.}

\item{as_table_list}{If TRUE, return the data model split into a list of tables rather than as one solid table.

loadDataModel}
}
\description{
Function reads in a CSV file containing the definitions of a given data
model definitions and returns either as one table or as a list of tables. By
default, this function returns the OMOP data model 'OMOP_CDM_v6_0_custom.csv' 
included in the extdata folder of the installed ROMOPOmics package, 
but similarly formatted tables can be used as well as long as they are packaged as CSVs.
}
