% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataModel.R
\name{loadDataModel}
\alias{loadDataModel}
\title{loadDataModel}
\usage{
loadDataModel(master_table_file, as_table_list = FALSE)
}
\description{
Function reads in "official" data model and returns as tibble (and list of tables).
}
\details{
loadDataModel
}
