% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelMask.R
\name{loadModelMasks}
\alias{loadModelMasks}
\title{loadModelMasks}
\usage{
loadModelMasks(mask_file_directory, data_model = loadDataModel())
}
\description{
Function reads in a "mask" table which includes three essential columns:
table, alias, and field. 'Alias' denotes the value being input, 'field'
denotes the data model's equivalent field name, and 'Table' denotes the
data model table where that field resides.
}
\details{
loadModelMasks
}
