% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mask.R
\name{read_mask}
\alias{read_mask}
\title{read_mask}
\usage{
read_mask(file_name)
}
\arguments{
\item{file_name}{File name to be read. Should specify a TSV/CSV file.

read_mask()}
}
\description{
Given a mask file TSV/CSV, function attempts to read it using fread and returns
an error message if it fails. Returns table with all applicable columns
appended.
}
