% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInputFiles.R
\name{readInputFiles}
\alias{readInputFiles}
\title{readInputFiles}
\usage{
readInputFiles(
  input_file = seq_input_files[1],
  data_model = dm,
  mask_table = msks$sequencing_mask
)
}
\arguments{
\item{input_file}{Name of a TSV file containing required alias column names.}

\item{data_model}{Data model being used, typically as a tibble returned by loadDataModel().}

\item{mask_table}{Mask contained in a tibble, typically as a tibble loaded by loadModelMask().

readInputFiles}
}
\description{
Function reads in TSV files designed with a given mask in mind, with rows
for each field and table combination and columns for input data entries.
Output is an "exhaustive" table including all fields and tables from the
specified data model, including unused tables and fields.
}
